Migrated MainActivity to HomeActivity to make room for Login page.
